Move folio_client to shared directory
